FitParameterSet moved from Fit library to Core (part II)
