+++ Fit/RootAdapter/GeneticMinimizer.cpp
@@ -13,7 +13,6 @@
 // ************************************************************************** //
 
 #include "GeneticMinimizer.h"
-#include "IFitParameter.h"
 #include "MinimizerConstants.h"
 #include "Math/GeneticMinimizer.h"
 #include "Parameter.h"
@@ -104,21 +103,6 @@ void GeneticMinimizer::setRandomSeed(int value)
 int GeneticMinimizer::randomSeed() const
 {
     return optionValue<int>(OptionNames::RandomSeed);
-}
-
-//! Sets minimizer parameter. Overload is required to check that parameter is properly limited.
-
-void GeneticMinimizer::setParameter(size_t index, const IFitParameter *par)
-{
-    if( !par->limits().isFixed() && !par->limits().isLimited()) {
-        std::ostringstream ostr;
-        ostr << "GeneticMinimizer::setParameter() -> Error! "
-             << "Genetic minimizer requires either fixed or "
-             << "limited AttLimits::limited(left,right) parameter. "
-             << " Parameter name '" << par->name() << "', limits:" << par->limits().toString();
-        throw std::runtime_error(ostr.str());
-    }
-    RootMinimizerAdapter::setParameter(index, par);
 }
 
 void GeneticMinimizer::setParameter(unsigned int index, const Fit::Parameter& par)
