+++ auto/Wrap/libBornAgainFit.py
@@ -2822,26 +2822,6 @@ Minimizer_swigregister = _libBornAgainFit.Minimizer_swigregister
 Minimizer_swigregister(Minimizer)
 
 
-class FitParameterSetIterator(object):
-
-    def __init__(self, fitParameters):
-        self.fitParameters = fitParameters
-        self.index = -1
-
-    def __iter__(self):
-        return self
-
-    def next(self):
-        self.index += 1
-        if self.index < self.fitParameters.size():
-            return self.fitParameters[self.index]
-        else:
-            raise StopIteration
-
-    def __next__(self):
-        return self.next()
-
-
 class ParametersIterator(object):
 
     def __init__(self, parameters):
